requirements.txt
# --- Core deep learning stack ---
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors

# --- Data processing & configuration ---
pandas
numpy
pyyaml
scikit-learn

# --- Large language model (Gemini API) ---
google-generativeai

# --- Progress bar & utilities ---
tqdm
pathlib
requests

# --- Visualization ---
matplotlib>=3.5.0